requirements.txt
setuptools==69.5.1
wheel==0.43.0
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.5.5.64
pandas==2.0.3
pillow==9.5.0
scikit-learn==1.3.2
seaborn==0.13.2
skorch==0.15.0
timm==0.5.4
torch==1.13.0
torchvision==0.14.0
xgboost==1.7.6
requests==2.31.0
Jinja2==3.1.2